nvim: Simplify config setup.
Move all the files into `plugin` folder. This ensures that everything is
loaded automatically, no `require(...)` is needed anymore.

For files in `plugin`, they are loaded alphabetically. To ensure that
plugins are loaded first, the file is prefixed with 1_. Other files
should not have any dependencies.

This also removes init.lua, which is now moved mostly to config.lua.